pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dedi-registry"
authors = [
  { name="Firefox2100", email="wangyunze16@gmail.com" },
]
description = "Decentralised Discovery Network Registry"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
dependencies = [
    "argon2-cffi~=25.1.0",
    "apscheduler~=3.11.0",
    "fastapi~=0.117.1",
    "itsdangerous~=2.2.0",
    "jinja2~=3.1.6",
    "jsonpatch~=1.33",
    "pyaml~=25.7.0",
    "pydantic~=2.11.7",
    "pydantic-settings~=2.11.0",
    "pymongo~=4.15.1",
    "python-multipart~=0.0.20",
    "redis~=6.4.0",
    "rich~=14.1.0",
    "typer~=0.19.2",
    "uvicorn~=0.37.0",
]
dynamic = [
    "version"
]

[project.urls]
Homepage = "https://github.com/Firefox2100/dedi-registry"
Repository = "https://github.com/Firefox2100/dedi-registry.git"
Issues = "https://github.com/Firefox2100/dedi-registry/issues"

[tool.setuptools.dynamic]
version = {attr = "dedi_registry.__version__"}


[project.scripts]
dedi-registry = "dedi_registry.app:main"
dedi-registry-cli = "dedi_registry.cli:cli_main"
